[JOBS] Minor improvement

+++ models/process_job.go
@@ -187,14 +187,17 @@ func (pj *ProcessJob) CloseInput() {
 
 // kill the job, runCommand will do the rest (I think)
 func (pj *ProcessJob) Kill() (e error) {
-	// Retrieve cmd
-	cmd := pj.cmd()
+	// Kill only if not completed
+	if <-completed[pj.JobId] != 0 {
+		// Retrieve cmd
+		cmd := pj.cmd()
 
-	// Kill the proc
-	e = cmd.Process.Kill()
+		// Kill the proc
+		e = cmd.Process.Kill()
 
-	if e != nil {
-		log.Error(e)
+		if e != nil {
+			log.Error(e)
+		}
 	}
 
 	return
